requirements.txt
flask==2.3.3
pymongo==4.6.1
python-dotenv==1.0.0
google-generativeai==0.3.1
werkzeug==2.3.7
markdown==3.5
bson==0.5.10 
Flask-Mail
itsdangerous